fixed problem where shooter would not launch ball
I think I fixed the problem where the shooter would not launch the ball.
The problem was that update() is only called in auton when timer.delay is
called, however, update is constantly called in teleop.(shooter works in
teleop) (see
operatorControl() function near the bottom of Robot.java) The update
function needs to be called for the shooter to start spinning the shooter
wheels or the intake wheels. I added timer.delay after
shooter.launch which probably solved the problem (not yet tested)